requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
PyMuPDF
weasyprint
jinja2
